Refactor and fix tests when no env vars given
